change fetch to align with server component fetching and refactor callback funcs
